+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/googleapis/enterprise-certificate-proxy
 
 go 1.19
+
+require github.com/google/go-pkcs11 v0.2.0
