--- internal/signer/linux/go.mod
@@ -1,5 +1,0 @@
-module signer
-
-go 1.19
-
-require github.com/google/go-pkcs11 v0.2.0
