+++ internal/signer/linux/pkcs11/pkcs11_test.go
@@ -1,0 +1,36 @@
+// Copyright 2023 Google LLC.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package pkcs11
+
+import (
+	"testing"
+)
+
+func TestParseHexString(t *testing.T) {
+	got, err := ParseHexString("0x1739427")
+	if err != nil {
+		t.Fatalf("ParseHexString error: %v", err)
+	}
+	want := uint32(0x1739427)
+	if got != want {
+		t.Errorf("Expected result is %v, got: %v", want, got)
+	}
+}
+
+func TestParseHexStringFailure(t *testing.T) {
+	_, err := ParseHexString("abcdefgh")
+	if err == nil {
+		t.Error("Expected error but got nil")
+	}
+}
